requirements.txt
fastapi[standard]
sqlmodel
passlib[bcrypt]
PyJWT
python-multipart
aiosqlite
dotenv
cloudinary
